CP Software/requirements.txt
flask >= 1.0.2
Jinja2 >= 2.8
MarkupSafe >= 0.23
Werkzeug >= 0.11.11
argparse >= 1.2.1
click >= 6.6
itsdangerous >= 0.24
# wsgiref>=0.1.2
quart >= 0.6.6
tox >= 3.2.1
pip >= 18.0
# python3.7.0
flask-compress >= 1.4.0
Flask-Static-Compress >= 1.0.2
jsmin >= 2.2.2
csscompressor >= 0.9.5
Flask-Compress >= 1.4.0
PyInstaller >= 3.4
py2app >= 0.18
Pillow >= 5.3.0
requests >= 2.21.0
mgrs >= 1.3.5